+++ flutter_sound/example/lib/loglevel/loglevel.dart
@@ -106,7 +106,8 @@ class _LogLevelState extends State<LogLevel> {
   @override
   Widget build(BuildContext context) {
     List<Widget> makeRadioButton() {
-      List<Widget> r ; r = [];
+      List<Widget> r;
+      r = [];
 
       r.add(
         Container(