+++ flutter_sound/example/lib/volume_control/volume_control.dart
@@ -1,0 +1,165 @@
+/*
+ * Copyright 2018, 2019, 2020 Dooboolab.
+ *
+ * This file is part of Flutter-Sound.
+ *
+ * Flutter-Sound is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License version 3 (LGPL-V3), as published by
+ * the Free Software Foundation.
+ *
+ * Flutter-Sound is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Flutter-Sound.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+import 'dart:async';
+import 'package:flutter/material.dart';
+import 'package:flutter_sound/flutter_sound.dart';
+
+/*
+ *
+ * This is a very simple example for Flutter Sound beginners,
+ * that show how to record, and then playback a file.
+ *
+ * This example is really basic.
+ *
+ */
+
+final _exampleAudioFilePathMP3 =
+    'https://file-examples-com.github.io/uploads/2017/11/file_example_MP3_700KB.mp3';
+
+///
+typedef Fn = void Function();
+
+/// Example app.
+class VolumeControl extends StatefulWidget {
+  @override
+  _VolumeControlState createState() => _VolumeControlState();
+}
+
+class _VolumeControlState extends State<VolumeControl> {
+  FlutterSoundPlayer? _mPlayer = FlutterSoundPlayer(logLevel: Level.debug);
+  bool _mPlayerIsInited = false;
+  double _mVolume = 100.0;
+
+  @override
+  void initState() {
+    super.initState();
+    _mPlayer!.openAudioSession().then((value) {
+      setState(() {
+        _mPlayerIsInited = true;
+      });
+    });
+  }
+
+  @override
+  void dispose() {
+    stopPlayer();
+    // Be careful : you must `close` the audio session when you have finished with it.
+    _mPlayer!.closeAudioSession();
+    _mPlayer = null;
+
+    super.dispose();
+  }
+
+  // -------  Here is the code to playback a remote file -----------------------
+
+  void play() async {
+    await _mPlayer!.startPlayer(
+        fromURI: _exampleAudioFilePathMP3,
+        codec: Codec.mp3,
+        whenFinished: () {
+          setState(() {});
+        });
+    setState(() {});
+  }
+
+  Future<void> stopPlayer() async {
+    if (_mPlayer != null) {
+      await _mPlayer!.stopPlayer();
+    }
+  }
+
+  Future<void> setVolume(double v) async // v is between 0.0 and 100.0
+  {
+    v = v > 100.0 ? 100.0 : v;
+    _mVolume = v;
+    setState(() {});
+    //await _mPlayer!.setVolume(v / 100, fadeDuration: Duration(milliseconds: 5000));
+    await _mPlayer!.setVolume(
+      v / 100,
+    );
+  }
+
+  // --------------------- UI -------------------
+
+  Fn? getPlaybackFn() {
+    if (!_mPlayerIsInited) {
+      return null;
+    }
+    return _mPlayer!.isStopped
+        ? play
+        : () {
+            stopPlayer().then((value) => setState(() {}));
+          };
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    Widget makeBody() {
+      //return Column(
+      //children: [
+      return Container(
+        margin: const EdgeInsets.all(3),
+        padding: const EdgeInsets.all(3),
+        height: 140,
+        width: double.infinity,
+        alignment: Alignment.center,
+        decoration: BoxDecoration(
+          color: Color(0xFFFAF0E6),
+          border: Border.all(
+            color: Colors.indigo,
+            width: 3,
+          ),
+        ),
+        child: Column(children: [
+          Row(children: [
+            ElevatedButton(
+              onPressed: getPlaybackFn(),
+              //color: Colors.white,
+              //disabledColor: Colors.grey,
+              child: Text(_mPlayer!.isPlaying ? 'Stop' : 'Play'),
+            ),
+            SizedBox(
+              width: 20,
+            ),
+            Text(_mPlayer!.isPlaying
+                ? 'Playback in progress'
+                : 'Player is stopped'),
+          ]),
+          Text('Volume:'),
+          Slider(
+              value: _mVolume,
+              min: 0.0,
+              max: 100.0,
+              onChanged: setVolume,
+              divisions: 100),
+        ]),
+        //),
+        //],
+      );
+    }
+
+    return Scaffold(
+      backgroundColor: Colors.blue,
+      appBar: AppBar(
+        title: const Text('Volume Control'),
+      ),
+      body: makeBody(),
+    );
+  }
+}