+++ tau_web/tau_web_example/typings/cordova-typings.d.ts
@@ -1,0 +1,2 @@
+
+/// <reference path="../.vscode/typings/cordova/cordova.d.ts"/>