+++ flutter_sound_platform_interface/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 8.2.0
+## 8.2.1
 
 - Please [look to this](https://dooboolab.github.io/flutter_sound/book/CHANGELOG.html)